feat: move post loading logic to separate file
